Testing the deconstructed HX block
Having the deconstructed HX block simply as two pipes separated by a metal wall with a HX topology showed that the Air/Air HX is indeed possible. The parameters need adjustment as the pressure drop of 6 bar within the HX is too high. Going to adjust the parameters and re-construct it back together into a single block.