Going back to make sure a HX component on its own is correct.
Tuning the standalone exploded co-current flow HX component. Set the input mass flow rates, temperatures and pressures according to specified scenario conditions. Both output temperatures and pressures seem within reasonable ranges, although there is a substantial pressure drop on both sides, which seems to be correlated with the drop in temperature.
Setting friction model to NoFriction removes the pressure drop in the pipes themselves, but leaving it to a nominal value of 0.005 results in an expected dP in both pipes. Also, adjusting the number of parallel pipes to 50 removed the crazy pressure values at the inlets. The model also now has mass flow sources and sinks on both sides instead of just one side.